Capstone project UI/Phase 1 UI/requirements.txt
flask
torch
torchvision
opencv-python
pillow
numpy